pyproject.toml
[tool.isort]
profile = 'black'
skip = ['env','env3.8','openly','simple_locations', 'typings']

[tool.black]
line-length = 180
target-version = ['py38']
include = '\.pyi?$'

[tool.poetry]
name = "djangostreetmap"
version = "0.1.0"
description = ""
authors = ["Joshua Brooks <josh.vdbroek@gmail.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.7"
osmium = "^3.2.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
mypy = "^0.910"
isort = "^5.9.3"
pre-commit = "^2.15.0"
pytest = "^6.2.5"
pytest-django = "^4.5.2"
requests-mock = "^1.9.3"
types-requests = "^2.27.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
